+++ packages/expo/src/specs/NativeClerkUserProfileView.android.ts
@@ -1,12 +1,11 @@
 import { requireNativeView } from 'expo';
-import type { ViewProps } from 'react-native';
+import type { NativeSyntheticEvent, ViewProps } from 'react-native';
 
 type ProfileEvent = Readonly<{ type: string }>;
-type NativeEvent<T> = Readonly<{ nativeEvent: T }>;
 
 interface NativeProps extends ViewProps {
   isDismissible?: boolean;
-  onProfileEvent?: (event: NativeEvent<ProfileEvent>) => void;
+  onProfileEvent?: (event: NativeSyntheticEvent<ProfileEvent>) => void;
 }
 
 export default requireNativeView<NativeProps>('ClerkUserProfileView');